+++ src/lightly_train/_task_models/dinov2_ltdetr_object_detection/transforms.py
@@ -18,6 +18,7 @@ from lightly_train._transforms.object_detection_transform import (
     ObjectDetectionTransformArgs,
 )
 from lightly_train._transforms.transform import (
+    MixUpArgs,
     NormalizeArgs,
     RandomFlipArgs,
     RandomIoUCropArgs,
@@ -110,6 +111,14 @@ class DINOv2LTDETRObjectDetectionScaleJitterArgs(ScaleJitterArgs):
     divisible_by: int | None = None
 
 
+class DINOv2LTDETRObjectDetectionMixUpArgs(MixUpArgs):
+    prob: float = 0.5
+    # Corresponds to the 4th epoch of the total training run.
+    step_start: int = 25_000  # TODO (Yutong 04/26): Update step_start and step_stop based on the actual number of training steps.
+    # Corresponds to the 4 + total_epochs // 2 epoch of the total training run.
+    step_stop: int = 250_000  # TODO (Yutong 04/26): Update step_start and step_stop based on the actual number of training steps.
+
+
 class DINOv2LTDETRObjectDetectionResizeArgs(ResizeArgs):
     height: int | Literal["auto"] = "auto"
     width: int | Literal["auto"] = "auto"
@@ -141,6 +150,7 @@ class DINOv2LTDETRObjectDetectionTrainTransformArgs(ObjectDetectionTransformArgs
     scale_jitter: DINOv2LTDETRObjectDetectionScaleJitterArgs | None = Field(
         default_factory=DINOv2LTDETRObjectDetectionScaleJitterArgs
     )
+    mixup: DINOv2LTDETRObjectDetectionMixUpArgs | None = None
     # We use the YOLO format internally for now.
     bbox_params: BboxParams = Field(
         default_factory=lambda: BboxParams(